requirements.txt
# Isntgram Essential Dependencies - Security Focused
# Created: July 27, 2025
# Purpose: Minimal, secure dependency list

# Core Framework
Flask>=3.1.1
Werkzeug>=3.0.6

# Database
SQLAlchemy>=2.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.7
Alembic>=1.13.2
psycopg2-binary>=2.9.9

# Authentication & Security
Flask-Login>=0.6.3
PyJWT>=2.10.1
Flask-WTF>=1.2.1
WTForms>=3.1.2
Flask-CORS>=4.0.2

# AWS Integration
boto3>=1.35.0

# Utilities
python-dotenv>=1.0.1

# Validation & Serialization
pydantic>=2.10.3
pydantic[email]>=2.10.3

# API Documentation & Rate Limiting
flask-limiter>=3.8.0
redis>=5.1.1
flasgger>=0.9.7.1

# Production Server
gunicorn>=23.0.0

# Testing & Coverage
pytest>=8.4.1
pytest-flask>=1.3.0
pytest-mock>=3.14.1
pytest-cov>=6.0.0
coverage>=7.6.0
factory-boy>=3.3.1
faker>=37.4.2
freezegun>=1.5.1
